+++ Cpp/binarysearch.h
@@ -2,8 +2,9 @@
 #include "common.h"
 
 namespace BinarySearch {
+	// find peak element in an array
 	int peakIndexInMountainArray(vector<int>& A) {
-		int l = 0, r = A.size();
+		int l = 0, r = (int)A.size();
 		while (l < r) {
 			int p = (l + r) >> 1;
 			if (A[p] > A[p + 1]) {
